# webapp/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  description: Mojito API Documentation
  title: Mojito API
  version: 0.111-SNAPSHOT
servers:
- description: Generated server url
  url: http://localhost:8080
paths:
...
  /api/branchStatistics:
    get:
      operationId: getBranchesOfRepository_1
      parameters:
      - in: query
        name: createdByUserName
        required: false
        schema:
          type: string
      - in: query
        name: branchId
        required: false
        schema:
          type: integer
          format: int64
      - in: query
        name: branchName
        required: false
        schema:
          type: string
      - in: query
        name: search
        required: false
        schema:
          type: string
      - in: query
        name: deleted
        required: false
        schema:
          type: boolean
      - in: query
        name: empty
        required: false
        schema:
          type: boolean
      - in: query
        name: totalCountLte
        required: false
        schema:
          type: integer
          format: int64
          default: 30000
      - in: query
        name: createdBefore
        required: false
        schema:
          type: string
          format: date-time
      - in: query
        name: createdAfter
        required: false
        schema:
          type: string
          format: date-time
      - description: Zero-based page index (0..N)
        in: query
        name: page
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - description: The size of the page to be returned
        in: query
        name: size
        required: false
        schema:
          type: integer
          default: 10
          minimum: 1
      - description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        in: query
        name: sort
        required: false
        schema:
          type: array
          default:
          - "id,DESC"
          items:
            type: string
      responses:
        "200":
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/PageBranchStatistic_BranchStatistic"
          description: OK
      summary: Get paginated Branch Statistics for a given set of parameters
      tags:
      - branch-statistic-ws
  /api/branchStatistics/{id}/branchTextUnitStatistics:
    get:
      operationId: getBranchTextUnitStatisticsOfBranch
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
      - description: Zero-based page index (0..N)
        in: query
        name: page
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - description: The size of the page to be returned
        in: query
        name: size
        required: false
        schema:
          type: integer
          default: 10
          minimum: 1
      - description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        in: query
        name: sort
        required: false
        schema:
          type: array
          default:
          - "id,DESC"
          items:
            type: string
      responses:
        "200":
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/PageBranchTextUnitStatistic_BranchTextUnitStatistic"
          description: OK
      summary: Get Branch Text Unit Statistics paginated for a specific Branch Statistic
      tags:
      - branch-statistic-ws
...
  /api/repositories/{repositoryId}/branches:
    delete:
      operationId: deleteBranch
      parameters:
      - in: path
        name: repositoryId
        required: true
        schema:
          type: integer
          format: int64
      - in: query
        name: branchId
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/PollableTask"
          description: OK
      summary: Delete a Branch asynchronously
      tags:
      - repository-ws
    get:
      operationId: getBranchesOfRepository
      parameters:
      - in: path
        name: repositoryId
        required: true
        schema:
          type: integer
          format: int64
      - in: query
        name: name
        required: false
        schema:
          type: string
      - in: query
        name: deleted
        required: false
        schema:
          type: boolean
      - in: query
        name: translated
        required: false
        schema:
          type: boolean
      - in: query
        name: createdBefore
        required: false
        schema:
          type: string
          format: date-time
      responses:
        "200":
          content:
            application/hal+json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch_BranchSummary"
          description: OK
      tags:
      - repository-ws
...
  /api/users:
    get:
      operationId: getUsers
      parameters:
      - in: query
        name: username
        required: false
        schema:
          type: string
      - in: query
        name: search
        required: false
        schema:
          type: string
      - description: Zero-based page index (0..N)
        in: query
        name: page
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - description: The size of the page to be returned
        in: query
        name: size
        required: false
        schema:
          type: integer
          default: 10
          minimum: 1
      - description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        in: query
        name: sort
        required: false
        schema:
          type: array
          default:
          - "username,ASC"
          items:
            type: string
      responses:
        "200":
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/PageUser"
          description: OK
      summary: Get paginated Users
      tags:
      - user-ws
    post:
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        "200":
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/User"
          description: OK
      tags:
      - user-ws
  /api/users/pw:
    post:
      operationId: changePassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordChangeRequest"
        required: true
      responses:
        "200":
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/User"
          description: OK
      summary: Update a user password
      tags:
      - user-ws
...
  /api/users/{userId}:
    delete:
      operationId: deleteUserByUserId
      parameters:
      - in: path
        name: userId
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
      tags:
      - user-ws
    patch:
      operationId: updateUserByUserId
      parameters:
      - in: path
        name: userId
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        "200":
          description: OK
      tags:
      - user-ws
...
components:
  schemas:
...
    Authority:
      type: object
      properties:
        authority:
          type: string
        createdDate:
          type: integer
          format: int64
          example: 1715699917000
        id:
          type: integer
          format: int64
        lastModifiedDate:
          type: integer
          format: int64
          example: 1715699917000
...
    BranchStatistic_BranchStatistic:
      type: object
      properties:
        branch:
          $ref: "#/components/schemas/Branch_BranchStatistic"
        branchTextUnitStatistics:
          type: array
          items:
            $ref: "#/components/schemas/BranchTextUnitStatistic_BranchStatistic"
          uniqueItems: true
        createdDate:
          type: integer
          format: int64
          example: 1715699917000
        forTranslationCount:
          type: integer
          format: int64
        id:
          type: integer
          format: int64
        totalCount:
          type: integer
          format: int64
...
    BranchTextUnitStatistic_BranchStatistic:
      type: object
      properties:
        forTranslationCount:
          type: integer
          format: int64
        id:
          type: integer
          format: int64
        tmTextUnit:
          $ref: "#/components/schemas/TMTextUnit_BranchStatistic"
        totalCount:
          type: integer
          format: int64
    BranchTextUnitStatistic_BranchTextUnitStatistic:
      type: object
      properties:
        forTranslationCount:
          type: integer
          format: int64
        id:
          type: integer
          format: int64
        tmTextUnit:
          $ref: "#/components/schemas/TMTextUnit_BranchTextUnitStatistic"
        totalCount:
          type: integer
          format: int64
    Branch_BranchStatistic:
      type: object
      properties:
        createdByUser:
          $ref: "#/components/schemas/User_BranchStatistic"
        createdDate:
          type: integer
          format: int64
          example: 1715699917000
        deleted:
          type: boolean
        id:
          type: integer
          format: int64
        name:
          type: string
        repository:
          $ref: "#/components/schemas/Repository_BranchStatistic"
        screenshots:
          type: array
          items:
            $ref: "#/components/schemas/Screenshot_BranchStatistic"
          uniqueItems: true
    Branch_BranchSummary:
      type: object
      properties:
        createdByUser:
          $ref: "#/components/schemas/User_BranchSummary"
        createdDate:
          type: integer
          format: int64
          example: 1715699917000
        deleted:
          type: boolean
        id:
          type: integer
          format: int64
        name:
          type: string
        repository:
          $ref: "#/components/schemas/Repository_BranchSummary"
        screenshots:
          type: array
          items:
            $ref: "#/components/schemas/Screenshot_BranchSummary"
          uniqueItems: true
...
    Locale_BranchStatistic:
      type: object
      properties:
        id:
          type: integer
          format: int64
...
    PageBranchStatistic_BranchStatistic:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/BranchStatistic_BranchStatistic"
        empty:
          type: boolean
        first:
          type: boolean
        last:
          type: boolean
        number:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject_BranchStatistic"
        size:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    PageBranchTextUnitStatistic_BranchTextUnitStatistic:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/BranchTextUnitStatistic_BranchTextUnitStatistic"
        empty:
          type: boolean
        first:
          type: boolean
        last:
          type: boolean
        number:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject_BranchTextUnitStatistic"
        size:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    PageUser:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/User"
        empty:
          type: boolean
        first:
          type: boolean
        last:
          type: boolean
        number:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        size:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        sort:
          $ref: "#/components/schemas/SortObject"
        unpaged:
          type: boolean
    PageableObject_BranchStatistic:
      type: object
    PageableObject_BranchTextUnitStatistic:
      type: object
...
    PasswordChangeRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
...
    PollableTask:
      type: object
      properties:
        allFinished:
          type: boolean
        createdByUser:
          $ref: "#/components/schemas/User"
        createdDate:
          type: integer
          format: int64
          example: 1715699917000
        errorMessage:
          type: string
        errorStack:
          type: string
        expectedSubTaskNumber:
          type: integer
          format: int32
        finishedDate:
          type: integer
          format: int64
          example: 1715699917000
        id:
          type: integer
          format: int64
        lastModifiedDate:
          type: integer
          format: int64
          example: 1715699917000
        message:
          type: string
        name:
          type: string
        subTasks:
          type: array
          items:
            $ref: "#/components/schemas/PollableTask"
          uniqueItems: true
...
    Repository_BranchStatistic:
      type: object
      properties:
        createdDate:
          type: integer
          format: int64
          example: 1715699917000
        id:
          type: integer
          format: int64
        manualScreenshotRun:
          $ref: "#/components/schemas/ScreenshotRun_BranchStatistic"
        name:
          type: string
        sourceLocale:
          $ref: "#/components/schemas/Locale_BranchStatistic"
    Repository_BranchSummary:
      type: object
      properties:
        createdDate:
          type: integer
          format: int64
          example: 1715699917000
        id:
          type: integer
          format: int64
        name:
          type: string
...
    ScreenshotRun_BranchStatistic:
      type: object
      properties:
        createdDate:
          type: integer
          format: int64
          example: 1715699917000
        id:
          type: integer
          format: int64
...
    ScreenshotTextUnit_BranchStatistic:
      type: object
      properties:
        id:
          type: integer
          format: int64
        tmTextUnit:
          $ref: "#/components/schemas/TMTextUnit_BranchStatistic"
...
    Screenshot_BranchStatistic:
      type: object
      properties:
        createdDate:
          type: integer
          format: int64
          example: 1715699917000
        id:
          type: integer
          format: int64
        src:
          type: string
        textUnits:
          type: array
          items:
            $ref: "#/components/schemas/ScreenshotTextUnit_BranchStatistic"
          uniqueItems: true
    Screenshot_BranchSummary:
      type: object
      properties:
        createdDate:
          type: integer
          format: int64
          example: 1715699917000
        id:
          type: integer
          format: int64
...
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...
    TMTextUnit_BranchStatistic:
      type: object
      properties:
        content:
          type: string
        createdDate:
          type: integer
          format: int64
          example: 1715699917000
        id:
          type: integer
          format: int64
        name:
          type: string
    TMTextUnit_BranchTextUnitStatistic:
      type: object
      properties:
        content:
          type: string
        id:
          type: integer
          format: int64
        name:
          type: string
...
    User:
      type: object
      properties:
        authorities:
          type: array
          items:
            $ref: "#/components/schemas/Authority"
          uniqueItems: true
        commonName:
          type: string
        createdDate:
          type: integer
          format: int64
          example: 1715699917000
        enabled:
          type: boolean
        givenName:
          type: string
        id:
          type: integer
          format: int64
        lastModifiedDate:
          type: integer
          format: int64
          example: 1715699917000
        partiallyCreated:
          type: boolean
        password:
          type: string
          writeOnly: true
        surname:
          type: string
        username:
          type: string
...
    User_BranchStatistic:
      type: object
      properties:
        commonName:
          type: string
        createdDate:
          type: integer
          format: int64
          example: 1715699917000
        id:
          type: integer
          format: int64
        username:
          type: string
    User_BranchSummary:
      type: object
      properties:
        commonName:
          type: string
        createdDate:
          type: integer
          format: int64
          example: 1715699917000
        id:
          type: integer
          format: int64
        username:
          type: string
...
  securitySchemes:
    cookieAuth:
      in: cookie
      name: JSESSIONID
      type: apiKey